+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Use multi-select in `task create` when no flags are used
 - Standardized language used in inputs
 - Update CI `yaml` file
+- Added `project delete`
 
 ## 2024-12-19 v0.6.18